Moved schema to client for easier imports
